+++ packages/core/src/index.ts
@@ -1,1 +1,4 @@
 export { Components, JSX } from './components';
+
+export * from './components/map-base-layers/map-base-layers.interfaces';
+export * from './components/map-overlays/map-overlays.interfaces';