+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * **dso-toolkit:** Smoke test voor `dso-toolkit` SCSS compile ([#1168](https://github.com/dso-toolkit/dso-toolkit/issues/1168))
+* **core + leaflet + react-leaflet:** MapControls Web Component, Leaflet plugins + styling, React Leaflet bindings ([#1022](https://github.com/dso-toolkit/dso-toolkit/issues/1022))
 
 ### Fixed
 * **sources:** Afbeeldingen in table schalen niet goed ([#1180](https://github.com/dso-toolkit/dso-toolkit/issues/1180))